requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]>=2.0,<3.0
alembic>=1.13,<2.0
asyncpg>=0.29.0,<1.0
psycopg2-binary>=2.9.9,<3.0
bcrypt>=4.0.0,<5.0.0
pydantic==2.9.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.20
httpx==0.25.2
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.1
email-validator==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

